% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_under_peaks.R
\name{area_under_peaks}
\alias{area_under_peaks}
\title{Get the Area Under Peaks}
\usage{
area_under_peaks(df)
}
\arguments{
\item{df}{a data.frame}
}
\value{
a data.frame with the coordinants on peaks
}
\description{
Get the area under peaks of local maximums
}
