% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chosen_peaks.R
\name{choose_peaks}
\alias{choose_peaks}
\title{Choose Peaks}
\usage{
choose_peaks(df, xs, dist_col = "cms", peaks_col = "drv1_smoothed3")
}
\arguments{
\item{df}{a data.frame}

\item{xs}{a data.frame}

\item{dist_col}{column name of the distance column}

\item{peaks_col}{column name of the smoothed data}
}
\value{
a data.frame with the coordinants on peaks
}
\description{
Pick the peaks that have the largest AUC,
}
