% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eseal_coords}
\alias{eseal_coords}
\title{Sample eseal analysis selections}
\format{A data frame with 55 rows and 5 variables:
\describe{
  \item{image_path}{price, in US dollars}
  \item{xmin}{min x coord}
  \item{xmax}{max x coord}
  \item{ymin}{min y coord}
  \item{ymax}{max y coord}
}}
\usage{
eseal_coords
}
\description{
The resulting data coords from using get_raster_coords()
}
\keyword{datasets}
