% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peaks.R
\name{get_peaks}
\alias{get_peaks}
\title{Get the Coordinants of Peaks}
\usage{
get_peaks(df)
}
\arguments{
\item{df}{a data.frame}
}
\value{
a data.frame with the coordinants on peaks
}
\description{
Get the coordinants of local maximums
}
