% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raster_coords.R
\name{get_raster_coords}
\alias{get_raster_coords}
\title{Get Analysis Selection from a list of Rasters}
\usage{
get_raster_coords(raster_obj)
}
\arguments{
\item{raster_obj}{list object with rasters read in with raster_obj = purrr::map(paths_to_rasters,raster)}
}
\value{
a data frame with with analysis selection x,y coordinants
}
\description{
Create a data frome with the analysis coordinants for each ultrasound image. this is an interactive plot that a user must click twice to select/create an alanysis selection comprised of coordinants along the x axis indicating the min and max that will be averaged/flattend by read ultra sound.  This is required before running read_ultrasound.
}
