% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_maxima.R
\name{local_maxima}
\alias{local_maxima}
\title{Find the Idecies of Local Maximum}
\usage{
local_maxima(x)
}
\arguments{
\item{x}{a numeric vector (timeseries)}
}
\value{
a vector of indecies of the local maxima
}
\description{
Find the local maxima (peaks) in a time series
}
