% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eseal_meta.R
\name{read_eseal_meta}
\alias{read_eseal_meta}
\title{Read Ultra Sound Metadata}
\usage{
read_eseal_meta(image_dir)
}
\arguments{
\item{image_dir}{path to a Session_* directory}
}
\value{
a data frame with ultrasound metadata
}
\description{
Read in a selection of metadata fields from a ultrasound Session_* folder
including image dimentions, y resolution (for converting to cm, image_dir, image_file,
imagename, givenname, familyname, createtime
}
