% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ultrasound.R
\name{read_ultrasound}
\alias{read_ultrasound}
\title{Read Ultra Sound and flatten  time series}
\usage{
read_ultrasound(path, xmin, xmax)
}
\arguments{
\item{path}{path to a png file}

\item{xmin}{the left x coordinant of the analysis selection}

\item{xmax}{the right x coordinant of the analysis selection}
}
\value{
a data frame with image path, y, and mean reflectivity
}
\description{
Read in a PNG ultrasound file made with a XX ultrasound divice in to r and
filter for the selected region of the image and collapse the time axis by
taking a mean value for every row of pixels.
}
